Replace mock in metamorph server tests
